requirements.txt
pytest
requests
PyYAML
pytest-html
jsonschema
